+++ proger210109/proger/App.js
@@ -14,6 +14,7 @@ import {
   View,
   Text,
   StatusBar,
+  Image,
 } from 'react-native';
 
 // import {
@@ -31,6 +32,10 @@ const App: () => React$Node = () => {
       <SafeAreaView style={{ flex: 1 }}>
         <View style={styles.mainViewStyle}>
           <Text>RN app</Text>
+          <Image
+            style={{ width: 200, height: 200, borderRadius: 50 }}
+            source={{ uri: 'https://avatars1.githubusercontent.com/u/31467572?s=460&u=2d7bff135af371274a2623e917be73f257752c17&v=4' }}
+          />
         </View>
       </SafeAreaView>
     </>
@@ -40,7 +45,6 @@ const styles = StyleSheet.create({
   mainViewStyle: {
     width: 200,
     height: 200,
-    backgroundColor: 'blue',
   },
 });